+++ packages/agent/CHANGELOG.md
@@ -1,4 +1,10 @@
-# @null-shot/agent
+# @nullshot/agent
+
+## 0.3.1
+
+### Patch Changes
+
+- c35a604: Fix dependencies
 
 ## 0.3.0